refactor: login
- extract to login package
- extract Event
