+++ presentation/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
             android:exported="true" />
 
         <activity
-            android:name=".activity.LoginActivity"
+            android:name=".activity.login.LoginActivity"
             android:exported="true">
 
             <intent-filter>
