+++ presentation/src/main/kotlin/team/sungbinland/sseukssak/activity/login/LoginViewModel.kt
@@ -5,7 +5,14 @@
  * Please see: https://github.com/sungbinland/sseukssak/blob/main/LICENSE.
  */
 
-package team.sungbinland.sseukssak.activity
+/*
+ * SseukSsak © 2022 Team Sungbinland. all rights reserved.
+ * SseukSsak license is under the MIT.
+ *
+ * Please see: https://github.com/sungbinland/sseukssak/blob/main/LICENSE.
+ */
+
+package team.sungbinland.sseukssak.activity.login
 
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.flow.MutableSharedFlow
@@ -18,21 +25,16 @@ class LoginViewModel : BaseViewModel() {
     val eventFlow = _eventFlow.asSharedFlow()
 
     fun kakaoLogin() {
-        event(Event.KakaoLogin(Unit))
+        event(Event.KakaoLogin)
     }
 
     fun skipLogin() {
-        event(Event.SkipLogin(Unit))
+        event(Event.SkipLogin)
     }
 
     private fun event(event: Event) {
         viewModelScope.launch {
             _eventFlow.emit(event)
         }
-    }
-
-    sealed class Event {
-        data class KakaoLogin(val p: Unit) : Event()
-        data class SkipLogin(val p: Unit) : Event()
     }
 }
